refactor: domain, view 패키지로 수정
