+++ src/main/java/racingCar/domain/StringUtil.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 import java.util.Arrays;
 import java.util.List;
