+++ src/main/java/racingCar/view/ResultView.java
@@ -1,4 +1,4 @@
-package racingCar.ui;
+package racingCar.view;
 
 import java.util.List;
